document-iq/document-iq-ml-pipeline/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
xgboost
python-dotenv